use netcdf logic from fv3core tests for physics/driver tests
