+++ fv3core/tests/savepoint/test_translate.py
@@ -1,401 +1,1 @@
-import logging
-import os
-from typing import Any, Dict
-
-import numpy as np
-import pytest
-
-import pace.dsl
-import pace.dsl.gt4py_utils as gt_utils
-import pace.util
-from pace.stencils.testing import SavepointCase, dataset_to_dict
-from pace.util.mpi import MPI
-from pace.util.testing import compare_scalar, success, success_array
-
-
-# this only matters for manually-added print statements
-np.set_printoptions(threshold=4096)
-
-OUTDIR = os.path.join(os.path.dirname(os.path.realpath(__file__)), "output")
-GPU_MAX_ERR = 1e-10
-GPU_NEAR_ZERO = 1e-15
-
-
-def platform():
-    in_docker = os.environ.get("IN_DOCKER", False)
-    return "docker" if in_docker else "metal"
-
-
-def sample_wherefail(
-    computed_data,
-    ref_data,
-    eps,
-    print_failures,
-    failure_stride,
-    test_name,
-    ignore_near_zero_errors,
-    near_zero,
-    xy_indices=False,
-):
-    found_indices = np.where(
-        np.logical_not(
-            success_array(
-                computed_data, ref_data, eps, ignore_near_zero_errors, near_zero
-            )
-        )
-    )
-    computed_failures = computed_data[found_indices]
-    reference_failures = ref_data[found_indices]
-
-    # List all errors
-    return_strings = []
-    bad_indices_count = len(found_indices[0])
-    # Determine worst result
-    worst_metric_err = 0.0
-    for b in range(bad_indices_count):
-        full_index = [f[b] for f in found_indices]
-        metric_err = compare_scalar(computed_failures[b], reference_failures[b])
-        abs_err = abs(computed_failures[b] - reference_failures[b])
-        if print_failures and b % failure_stride == 0:
-            return_strings.append(
-                f"index: {full_index}, computed {computed_failures[b]}, "
-                f"reference {reference_failures[b]}, "
-                f"absolute diff {abs_err:.3e}, "
-                f"metric diff: {metric_err:.3e}"
-            )
-        if np.isnan(metric_err) or (metric_err > worst_metric_err):
-            worst_metric_err = metric_err
-            worst_full_idx = full_index
-            worst_abs_err = abs_err
-            computed_worst = computed_failures[b]
-            reference_worst = reference_failures[b]
-    # Summary and worst result
-    fullcount = len(ref_data.flatten())
-    return_strings.append(
-        f"Failed count: {bad_indices_count}/{fullcount} "
-        f"({round(100.0 * (bad_indices_count / fullcount), 2)}%),\n"
-        f"Worst failed index {worst_full_idx}\n"
-        f"\tcomputed:{computed_worst}\n"
-        f"\treference: {reference_worst}\n"
-        f"\tabsolute diff: {worst_abs_err:.3e}\n"
-        f"\tmetric diff: {worst_metric_err:.3e}\n"
-    )
-
-    if xy_indices:
-        if len(computed_data.shape) == 3:
-            axis = 2
-            any = np.any
-        elif len(computed_data.shape) == 4:
-            axis = (2, 3)
-            any = np.any
-        else:
-            axis = None
-
-            def any(array, axis):
-                return array
-
-        found_xy_indices = np.where(
-            any(
-                np.logical_not(
-                    success_array(
-                        computed_data, ref_data, eps, ignore_near_zero_errors, near_zero
-                    )
-                ),
-                axis=axis,
-            )
-        )
-
-        return_strings.append(
-            "failed horizontal indices:" + str(list(zip(*found_xy_indices)))
-        )
-
-    return "\n".join(return_strings)
-
-
-def process_override(threshold_overrides, testobj, test_name, backend):
-    override = threshold_overrides.get(test_name, None)
-    # NOTE (jdahm): Temporary until Jenkins is updated
-    if backend.startswith("gtc:"):
-        backend = backend.replace("gtc:", "")
-    if override is not None:
-        for spec in override:
-            if "platform" not in spec:
-                spec["platform"] = platform()
-            if "backend" not in spec:
-                spec["backend"] = backend
-        matches = [
-            spec
-            for spec in override
-            if spec["backend"] == backend and spec["platform"] == platform()
-        ]
-        if len(matches) == 1:
-            match = matches[0]
-            if "max_error" in match:
-                testobj.max_error = float(match["max_error"])
-            if "near_zero" in match:
-                testobj.near_zero = float(match["near_zero"])
-            if "ignore_near_zero_errors" in match:
-                parsed_ignore_zero = match["ignore_near_zero_errors"]
-                if isinstance(parsed_ignore_zero, list):
-                    testobj.ignore_near_zero_errors.update(
-                        {field: True for field in match["ignore_near_zero_errors"]}
-                    )
-                elif isinstance(parsed_ignore_zero, dict):
-                    for key in parsed_ignore_zero.keys():
-                        testobj.ignore_near_zero_errors[key] = {}
-                        testobj.ignore_near_zero_errors[key]["near_zero"] = float(
-                            parsed_ignore_zero[key]
-                        )
-                    if "all_other_near_zero" in match:
-                        for key in testobj.out_vars.keys():
-                            if key not in testobj.ignore_near_zero_errors:
-                                testobj.ignore_near_zero_errors[key] = {}
-                                testobj.ignore_near_zero_errors[key][
-                                    "near_zero"
-                                ] = float(match["all_other_near_zero"])
-
-                else:
-                    raise TypeError(
-                        "ignore_near_zero_errors is either a list or a dict"
-                    )
-        elif len(matches) > 1:
-            raise Exception(
-                "misconfigured threshold overrides file, more than 1 specification for "
-                + test_name
-                + " with backend="
-                + backend
-                + ", platform="
-                + platform()
-            )
-
-
-@pytest.mark.sequential
-@pytest.mark.skipif(
-    MPI is not None and MPI.COMM_WORLD.Get_size() > 1,
-    reason="Running in parallel with mpi",
-)
-def test_sequential_savepoint(
-    case: SavepointCase,
-    backend,
-    print_failures,
-    failure_stride,
-    subtests,
-    caplog,
-    threshold_overrides,
-    xy_indices=True,
-):
-    caplog.set_level(logging.DEBUG, logger="fv3core")
-    if case.testobj is None:
-        pytest.xfail(
-            f"no translate object available for savepoint {case.savepoint_name}"
-        )
-    stencil_config = pace.dsl.StencilConfig(backend=backend)
-    # Reduce error threshold for GPU
-    if stencil_config.is_gpu_backend:
-        case.testobj.max_error = max(case.testobj.max_error, GPU_MAX_ERR)
-        case.testobj.near_zero = max(case.testobj.near_zero, GPU_NEAR_ZERO)
-    if threshold_overrides is not None:
-        process_override(
-            threshold_overrides, case.testobj, case.savepoint_name, backend
-        )
-    input_data = dataset_to_dict(case.ds_in)
-    input_names = (
-        case.testobj.serialnames(case.testobj.in_vars["data_vars"])
-        + case.testobj.in_vars["parameters"]
-    )
-    input_data = {name: input_data[name] for name in input_names}
-    # run python version of functionality
-    output = case.testobj.compute(input_data)
-    failing_names = []
-    passing_names = []
-    all_ref_data = dataset_to_dict(case.ds_out)
-    for varname in case.testobj.serialnames(case.testobj.out_vars):
-        ignore_near_zero = case.testobj.ignore_near_zero_errors.get(varname, False)
-        ref_data = all_ref_data[varname]
-        if hasattr(case.testobj, "subset_output"):
-            ref_data = case.testobj.subset_output(varname, ref_data)
-        with subtests.test(varname=varname):
-            failing_names.append(varname)
-            output_data = gt_utils.asarray(output[varname])
-            assert success(
-                output_data,
-                ref_data,
-                case.testobj.max_error,
-                ignore_near_zero,
-                case.testobj.near_zero,
-            ), sample_wherefail(
-                output_data,
-                ref_data,
-                case.testobj.max_error,
-                print_failures,
-                failure_stride,
-                case.savepoint_name,
-                ignore_near_zero_errors=ignore_near_zero,
-                near_zero=case.testobj.near_zero,
-                xy_indices=xy_indices,
-            )
-            passing_names.append(failing_names.pop())
-    if len(failing_names) > 0:
-        out_filename = os.path.join(OUTDIR, f"{case.savepoint_name}.nc")
-        try:
-            save_netcdf(
-                case.testobj,
-                [input_data],
-                [output],
-                ref_data,
-                failing_names,
-                out_filename,
-            )
-        except Exception as error:
-            print(f"TestSequential SaveNetCDF Error: {error}")
-    assert failing_names == [], f"only the following variables passed: {passing_names}"
-    assert len(passing_names) > 0, "No tests passed"
-
-
-def state_from_savepoint(serializer, savepoint, name_to_std_name):
-    properties = pace.util.fortran_info.properties_by_std_name
-    origin = gt_utils.origin
-    state = {}
-    for name, std_name in name_to_std_name.items():
-        array = serializer.read(name, savepoint)
-        extent = tuple(np.asarray(array.shape) - 2 * np.asarray(origin))
-        state["air_temperature"] = pace.util.Quantity(
-            array,
-            dims=reversed(properties["air_temperature"]["dims"]),
-            units=properties["air_temperature"]["units"],
-            origin=origin,
-            extent=extent,
-        )
-    return state
-
-
-def get_communicator(comm, layout):
-    partitioner = pace.util.CubedSpherePartitioner(pace.util.TilePartitioner(layout))
-    communicator = pace.util.CubedSphereCommunicator(comm, partitioner)
-    return communicator
-
-
-@pytest.mark.parallel
-@pytest.mark.skipif(
-    MPI is None or MPI.COMM_WORLD.Get_size() == 1,
-    reason="Not running in parallel with mpi",
-)
-def test_parallel_savepoint(
-    case: SavepointCase,
-    backend,
-    print_failures,
-    failure_stride,
-    subtests,
-    caplog,
-    threshold_overrides,
-    compute_grid,
-    xy_indices=True,
-):
-    layout = (
-        int((MPI.COMM_WORLD.Get_size() // 6) ** 0.5),
-        int((MPI.COMM_WORLD.Get_size() // 6) ** 0.5),
-    )
-    communicator = get_communicator(MPI.COMM_WORLD, layout)
-    caplog.set_level(logging.DEBUG, logger="fv3core")
-    if case.testobj is None:
-        pytest.xfail(
-            f"no translate object available for savepoint {case.savepoint_name}"
-        )
-    stencil_config = pace.dsl.StencilConfig(backend=backend)
-    # Increase minimum error threshold for GPU
-    if stencil_config.is_gpu_backend:
-        case.testobj.max_error = max(case.testobj.max_error, GPU_MAX_ERR)
-        case.testobj.near_zero = max(case.testobj.near_zero, GPU_NEAR_ZERO)
-    if threshold_overrides is not None:
-        process_override(
-            threshold_overrides, case.testobj, case.savepoint_name, backend
-        )
-    if compute_grid and not case.testobj.compute_grid_option:
-        pytest.xfail(f"compute_grid option not used for test {case.savepoint_name}")
-    input_data = dataset_to_dict(case.ds_in)
-    # run python version of functionality
-    output = case.testobj.compute_parallel(input_data, communicator)
-    out_vars = set(case.testobj.outputs.keys())
-    out_vars.update(list(case.testobj._base.out_vars.keys()))
-    failing_names = []
-    passing_names = []
-    ref_data: Dict[str, Any] = {}
-    all_ref_data = dataset_to_dict(case.ds_out)
-    for varname in out_vars:
-        ref_data[varname] = []
-        new_ref_data = all_ref_data[varname]
-        if hasattr(case.testobj, "subset_output"):
-            new_ref_data = case.testobj.subset_output(varname, new_ref_data)
-        ref_data[varname].append(new_ref_data)
-        ignore_near_zero = case.testobj.ignore_near_zero_errors.get(varname, False)
-        with subtests.test(varname=varname):
-            failing_names.append(varname)
-            output_data = gt_utils.asarray(output[varname])
-            assert success(
-                output_data,
-                ref_data[varname][0],
-                case.testobj.max_error,
-                ignore_near_zero,
-                case.testobj.near_zero,
-            ), sample_wherefail(
-                output_data,
-                ref_data[varname][0],
-                case.testobj.max_error,
-                print_failures,
-                failure_stride,
-                case.savepoint_name,
-                ignore_near_zero,
-                case.testobj.near_zero,
-                xy_indices,
-            )
-            passing_names.append(failing_names.pop())
-    if len(failing_names) > 0:
-        out_filename = os.path.join(
-            OUTDIR, f"{case.savepoint_name}-{case.grid[0].rank}.nc"
-        )
-        try:
-            save_netcdf(
-                case.testobj,
-                [input_data],
-                [output],
-                ref_data,
-                failing_names,
-                out_filename,
-            )
-        except Exception as error:
-            print(f"TestParallel SaveNetCDF Error: {error}")
-    assert failing_names == [], f"only the following variables passed: {passing_names}"
-    assert len(passing_names) > 0, "No tests passed"
-
-
-def save_netcdf(
-    testobj, inputs_list, output_list, ref_data, failing_names, out_filename
-):
-    import xarray as xr
-
-    data_vars = {}
-    for i, varname in enumerate(failing_names):
-        dims = [dim_name + f"_{i}" for dim_name in testobj.outputs[varname]["dims"]]
-        attrs = {"units": testobj.outputs[varname]["units"]}
-        try:
-            data_vars[f"{varname}_in"] = xr.DataArray(
-                np.stack([in_data[varname] for in_data in inputs_list]),
-                dims=("rank",) + tuple(dims),
-                attrs=attrs,
-            )
-        except KeyError as error:
-            print(f"No input data found for {error}")
-        data_vars[f"{varname}_ref"] = xr.DataArray(
-            np.stack(ref_data[varname]), dims=("rank",) + tuple(dims), attrs=attrs
-        )
-        data_vars[f"{varname}_out"] = xr.DataArray(
-            np.stack([output[varname] for output in output_list]),
-            dims=("rank",) + tuple(dims),
-            attrs=attrs,
-        )
-        data_vars[f"{varname}_error"] = (
-            data_vars[f"{varname}_ref"] - data_vars[f"{varname}_out"]
-        )
-        data_vars[f"{varname}_error"].attrs = attrs
-    print(f"File saved to {out_filename}")
-    xr.Dataset(data_vars=data_vars).to_netcdf(out_filename)
+from pace.stencils.testing.test_translate import *  # noqa: F403,F401
