+++ fv3gfs-physics/tests/conftest.py
@@ -15,9 +15,8 @@ def pytest_addoption(parser):
     parser.addoption("--print_failures", action="store_true")
     parser.addoption("--failure_stride", action="store", default=1)
     parser.addoption("--data_path", action="store", default="./")
-    parser.addoption("--python_regression", action="store_true")
     parser.addoption("--threshold_overrides_file", action="store", default=None)
-    parser.addoption("--print_domains", action="store_true")
+    parser.addoption("--compute_grid", action="store_true")
 
 
 def pytest_configure(config):
