+++ fv3gfs-physics/tests/savepoint/translate/translate_driver.py
@@ -16,8 +16,8 @@ enable_selective_validation()
 
 
 class TranslateDriver(TranslateFVDynamics):
-    def __init__(self, grids, namelist, stencil_factory):
-        super().__init__(grids[0], namelist, stencil_factory)
+    def __init__(self, grid, namelist, stencil_factory):
+        super().__init__(grid, namelist, stencil_factory)
         self.namelist: Namelist = namelist
         self.stencil_factory = stencil_factory
         self.stencil_config = self.stencil_factory.config
